+++ apps/studio/src/features/editing-experience/components/form-builder/renderers/controls/JsonFormsImageControl.tsx
@@ -1,32 +1,22 @@
 import type { ControlProps, RankedTester } from "@jsonforms/core"
-import { useState } from "react"
-import {
-  Button,
-  IconButton,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalFooter,
-  ModalHeader,
-  ModalOverlay,
-  Text,
-  useDisclosure,
-} from "@chakra-ui/react"
+import { Box, FormControl } from "@chakra-ui/react"
 import { and, isStringControl, rankWith, schemaMatches } from "@jsonforms/core"
 import { withJsonFormsControlProps } from "@jsonforms/react"
-import { BiTrash } from "react-icons/bi"
+import { FormErrorMessage, FormLabel } from "@opengovsg/design-system-react"
 import { z } from "zod"
 
-import { ImageUploadInfobox } from "~/components/ImageUploadInfobox"
 import { FileAttachment } from "~/components/PageEditor/FileAttachment"
 import { JSON_FORMS_RANKING } from "~/constants/formBuilder"
+import { env } from "~/env.mjs"
 import { useQueryParse } from "~/hooks/useQueryParse"
+import { useImage } from "../../hooks/useImage"
 import {
   IMAGE_UPLOAD_ACCEPTED_MIME_TYPE_MAPPING,
   MAX_IMG_FILE_SIZE_BYTES,
 } from "./constants"
+import { getCustomErrorMessage } from "./utils"
 
+const assetsBaseUrl = `https://${env.NEXT_PUBLIC_S3_ASSETS_DOMAIN_NAME}`
 export const jsonFormsImageControlTester: RankedTester = rankWith(
   JSON_FORMS_RANKING.ImageControl,
   and(
@@ -47,66 +37,32 @@ export function JsonFormsImageControl({
   handleChange,
   path,
   required,
-  data,
+  errors,
+  description,
 }: JsonFormsImageControlProps) {
   const { siteId } = useQueryParse(editSiteSchema)
-  const { isOpen, onClose, onOpen } = useDisclosure()
-  const [href, setHref] = useState("")
+  const { handleImage } = useImage({})
 
   return (
-    <>
-      <ImageUploadInfobox
-        description={"Upload an image"}
-        label={label}
-        onClick={onOpen}
-        required={required}
-      >
-        {!!data && (
-          <>
-            <Text noOfLines={1}>{data.split("/").pop()}</Text>
-            <IconButton
-              size="xs"
-              variant="clear"
-              colorScheme="critical"
-              aria-label="Remove file"
-              icon={<BiTrash />}
-              onClick={() => handleChange(path, undefined)}
-            />
-          </>
-        )}
-      </ImageUploadInfobox>
-
-      <Modal isOpen={isOpen} onClose={onClose}>
-        <ModalOverlay />
-        <ModalContent>
-          <ModalHeader mr="3.5rem">Add an image</ModalHeader>
-          <ModalCloseButton size="lg" />
-
-          <ModalBody>
-            <FileAttachment
-              maxSizeInBytes={MAX_IMG_FILE_SIZE_BYTES}
-              acceptedFileTypes={IMAGE_UPLOAD_ACCEPTED_MIME_TYPE_MAPPING}
-              siteId={siteId}
-              setHref={(image) => setHref(image ?? "")}
-            />
-          </ModalBody>
-
-          <ModalFooter>
-            <Button
-              variant="solid"
-              onClick={() => {
-                handleChange(path, href)
-                onClose()
-              }}
-              isDisabled={!href}
-              type="submit"
-            >
-              Add an image
-            </Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
-    </>
+    <Box as={FormControl} isRequired={required} isInvalid={!!errors}>
+      <FormLabel description={description}>{label}</FormLabel>
+      <FileAttachment
+        maxSizeInBytes={MAX_IMG_FILE_SIZE_BYTES}
+        acceptedFileTypes={IMAGE_UPLOAD_ACCEPTED_MIME_TYPE_MAPPING}
+        siteId={siteId}
+        setHref={(src) =>
+          src &&
+          handleImage(`${assetsBaseUrl}${src}`).then((src) =>
+            handleChange(path, src),
+          )
+        }
+      />
+      {!!errors && (
+        <FormErrorMessage>
+          {label} {getCustomErrorMessage(errors)}
+        </FormErrorMessage>
+      )}
+    </Box>
   )
 }
